// Reading Data/my-app/src/app/Model/Employee.ts
export interface Employee {
    id: number;
    name: string;
    gender: string;
    email?: string;
    phoneNumber?: number;
    contactPreference: string;
    dateOfBirth: Date;
    department: string;
    isActive: boolean;
    photoPath?: string;
    password?:string;
    conformPassword?:string;
}